added casts for floating-point operations in util.Geom

+++ jhotdraw6/src/org/jhotdraw/util/Geom.java
@@ -47,8 +47,9 @@ public class Geom {
 			return (Math.abs(py - y1) < 3);
 		}
 
-		a = (y1 - y2) / (x1 - x2);
-		b = y1 - a * x1;
+		// casts to avoid integer arithmetic instead of floating-point
+		a = (double)(y1 - y2) / (double)(x1 - x2);
+		b = (double)y1 - a * (double)x1;
 		x = (py - b) / a;
 		y = a * px + b;
